Introduce dotenv package for finding environment files
