+++ dev/scripts/src/scripts/generateMnemonic.ts
@@ -12,32 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { loadEnv } from "@prosopo/cli";
+import { loadEnv } from "@prosopo/dotenv";
 import { LogLevel, getLogger } from "@prosopo/common";
 import { generateMnemonic } from "@prosopo/contract";
 import { updateEnvFile } from "../setup/index.js";
 
 loadEnv();
 const logger = getLogger(
-	process.env.PROSOPO_LOG_LEVEL || LogLevel.enum.info,
-	"generateMnemonic",
+  process.env.PROSOPO_LOG_LEVEL || LogLevel.enum.info,
+  "generateMnemonic",
 );
 
 async function mnemonic(addToEnv: boolean) {
-	const [mnemonic, address] = await generateMnemonic();
-	logger.info(`Address: ${address}`);
-	logger.info(`Mnemonic: ${mnemonic}`);
-	if (addToEnv) {
-		await updateEnvFile({
-			PROSOPO_PROVIDER_MNEMONIC: `"${mnemonic}"`,
-			PROSOPO_PROVIDER_ADDRESS: address,
-		});
-	}
+  const [mnemonic, address] = await generateMnemonic();
+  logger.info(`Address: ${address}`);
+  logger.info(`Mnemonic: ${mnemonic}`);
+  if (addToEnv) {
+    await updateEnvFile({
+      PROSOPO_PROVIDER_MNEMONIC: `"${mnemonic}"`,
+      PROSOPO_PROVIDER_ADDRESS: address,
+    });
+  }
 }
 
 mnemonic(process.argv.includes("--env"))
-	.then(() => process.exit(0))
-	.catch((e) => {
-		console.error(e);
-		process.exit(1);
-	});
+  .then(() => process.exit(0))
+  .catch((e) => {
+    console.error(e);
+    process.exit(1);
+  });
