+++ packages/cli/src/index.ts
@@ -17,4 +17,3 @@ export * from "./start.js";
 export * from "./process.env.js";
 export * from "./files.js";
 export { default as defaultConfig } from "./prosopo.config.js";
-export * from "./env.js";
