+++ packages/cli/src/reloader.ts
@@ -18,62 +18,62 @@ import { LogLevel, getLogger } from "@prosopo/common";
 import { ProviderEnvironment } from "@prosopo/env";
 import type { ProsopoConfigOutput } from "@prosopo/types";
 import type { AwaitedProcessedArgs } from "./argv.js";
-import { loadEnv } from "./env.js";
+import { loadEnv } from "@prosopo/dotenv";
 import { start } from "./start.js";
 
 const log = getLogger(LogLevel.enum.info, "CLI");
 
 export default class ReloadingAPI {
-	// biome-ignore lint/suspicious/noExplicitAny: TODO fix
-	private _envWatcher: any;
-	private _envPath: string;
-	private _config: ProsopoConfigOutput;
-	private _pair: KeyringPair;
-	private _processedArgs: AwaitedProcessedArgs;
-	private api: Server | undefined;
-	private _restarting: boolean;
+  // biome-ignore lint/suspicious/noExplicitAny: TODO fix
+  private _envWatcher: any;
+  private _envPath: string;
+  private _config: ProsopoConfigOutput;
+  private _pair: KeyringPair;
+  private _processedArgs: AwaitedProcessedArgs;
+  private api: Server | undefined;
+  private _restarting: boolean;
 
-	constructor(
-		envPath: string,
-		config: ProsopoConfigOutput,
-		pair: KeyringPair,
-		processedArgs: AwaitedProcessedArgs,
-	) {
-		this._envPath = envPath;
-		this._config = config;
-		this._pair = pair;
-		this._processedArgs = processedArgs;
-		this._restarting = false;
-	}
+  constructor(
+    envPath: string,
+    config: ProsopoConfigOutput,
+    pair: KeyringPair,
+    processedArgs: AwaitedProcessedArgs,
+  ) {
+    this._envPath = envPath;
+    this._config = config;
+    this._pair = pair;
+    this._processedArgs = processedArgs;
+    this._restarting = false;
+  }
 
-	public async start() {
-		log.info("Starting API");
-		this._envWatcher = await this._watchEnv();
-		loadEnv();
-		const env = new ProviderEnvironment(this._config, this._pair);
-		await env.isReady();
-		this.api = await start(env, !!this._processedArgs.adminApi);
-	}
+  public async start() {
+    log.info("Starting API");
+    this._envWatcher = await this._watchEnv();
+    loadEnv();
+    const env = new ProviderEnvironment(this._config, this._pair);
+    await env.isReady();
+    this.api = await start(env, !!this._processedArgs.adminApi);
+  }
 
-	public async stop() {
-		log.info("Stopping API");
-		return new Promise((resolve) => {
-			if (this.api) {
-				this.api.close(resolve);
-			}
-		});
-	}
+  public async stop() {
+    log.info("Stopping API");
+    return new Promise((resolve) => {
+      if (this.api) {
+        this.api.close(resolve);
+      }
+    });
+  }
 
-	private async _watchEnv() {
-		return fs.watchFile(this._envPath, async () => {
-			log.info(`env file change detected. Restarting: ${this._restarting}`);
-			if (!this._restarting) {
-				this._restarting = true;
-				await this.stop();
-				loadEnv();
-				await this.start();
-				this._restarting = false;
-			}
-		});
-	}
+  private async _watchEnv() {
+    return fs.watchFile(this._envPath, async () => {
+      log.info(`env file change detected. Restarting: ${this._restarting}`);
+      if (!this._restarting) {
+        this._restarting = true;
+        await this.stop();
+        loadEnv();
+        await this.start();
+        this._restarting = false;
+      }
+    });
+  }
 }
