+++ packages/cli/tsconfig.cjs.json
@@ -16,6 +16,9 @@
 			"path": "../contract/tsconfig.cjs.json"
 		},
 		{
+			"path": "../dotenv/tsconfig.cjs.json"
+		},
+		{
 			"path": "../env/tsconfig.cjs.json"
 		},
 		{
