+++ packages/util/package.json
@@ -33,14 +33,13 @@
 	"dependencies": {
 		"@noble/hashes": "^1.3.3",
 		"@prosopo/config": "2.0.1",
-		"@prosopo/util": "2.0.1",
+		"dotenv": "^16.0.1",
 		"lodash": "^4.17.21",
 		"seedrandom": "^3.0.5"
 	},
 	"devDependencies": {
 		"@types/lodash": "^4.14.198",
 		"@types/seedrandom": "^3.0.5",
-		"dotenv": "^16.0.1",
 		"tslib": "2.6.2",
 		"typescript": "5.1.6",
 		"vitest": "^1.3.1"
